spidershim: Auto-generate the names of tests
This simplifies the process of adding a test by reading tests.gyp and
getting the names of all of the targets defined in that file as the
names of the tests to run, so there is no need to manually change
run-tests.sh any more.

This also rewrites run-tests.sh in Python now that part of it requires
Python.