spidershim: Avoid using a Context in HandleScope's constructor
It turns out that storing a JSContext* in the Context object wasn't a
good idea.  This patch moves JSContext* to Isolate, next to the
JSRuntime* that's managed there.  Then it fixes the setup for Isolate
persistents and eternals so that their RootStores are setup and torn
down at the right times.  Then it stores the Context's global object
in a Persistent<Object> instead of a Local<Object>, which removes the
need for a HandleScope to be on the stack when creating a Context.
After this, HandleScope is disentangled from Context.

With this, it's no longer needed for there to be a HandleScope on the
stack when creating a Context.  We also avoid creating a dummy Context
each time a HandleScope is created.