spidershim: Support rooting the native function return values to the parent HandleScope
V8 stores such handles in a separate location that is rooted, but
with our HandleScope implementation, we would need to root the handle
against the parent HandleScope, if one exists.

Without this, the test in this commit would UAF because the Local
created in the callback would be bound to the HandleScope within that
frame.