spidershim: rework how the compartment stack is managed
it doesn't work to store the old compartment in a context because a context can
be reentered in which case we'd say the previous context was our selves, and
then be unable to unwind the stack.  However we already have a stack of entered
contexts, and those should directly map to compartments.  So it should be
possible to use the context stack to get the compartment we want to leave to.