Revert "spidershim: rework how the compartment stack is managed"
This reverts commit fa48fdf9f79bb4468c2a2051d7ea00aeb8dd34c0.

This was actually breaking SpiderShim.TryCatchMixedNesting.  The
scenario described in https://github.com/mozilla/spidernode/issues/174
was happening in that test.

This commit was violating the contract of the JS_Enter/LeaveCompartment
APIs in that it was breaking the tree order of the calls.  If we need
this for some reason, then we need to find an alternative solution.